Move logic for editing venues into the venue detail component instead of venue card.
